refactor: move search string formatter into utils/search
