+++ src/utils/formatters.ts
@@ -1,25 +1,5 @@
 // Helper function to format component names
 
-export type SearchStringSplit = {
-  normalizedValue: string;
-  rawTokens: string[];
-};
-
-export function formatSearchString(value: string): SearchStringSplit | undefined {
-  const normalizedValue = value.trim().toLowerCase();
-
-  if (!normalizedValue) {
-    return undefined;
-  }
-
-  const rawTokens = normalizedValue
-    .split(/[^a-z0-9]+/)
-    .map((token) => token.trim())
-    .filter(Boolean);
-
-  return { normalizedValue, rawTokens };
-}
-
 export function formatComponentName(name: string): string {
   return name
     .split("-")
@@ -37,6 +17,3 @@ export function formatDisplayName(name: string): string {
     })
     .join(" ");
 }
-
-
-
